requirements.txt
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
ghp-import==2.0.1
idna==3.2
importlib-metadata==4.8.1
Jinja2==3.0.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
mkdocs-material==7.2.8
mkdocs-material-extensions==1.0.3
packaging==21.0
Pygments==2.10.0
pymdown-extensions==8.2
pyparsing==2.4.7
pyquery==1.4.3
python-dateutil==2.8.2
python-frontmatter==1.0.0
PyYAML==5.4.1
pyyaml_env_tag==0.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==2.1.5
webencodings==0.5.1
zipp==3.5.0
